CLEANUP: Modified the email handler to point to the path /update from the CMP host domain. This avoids combing user interface functionality with the response from the an email link.
The handler (cmp.handlerUpdate) that processes the email link now creates the update operation using the parameters in the query string. The changes to swanop now ensure a new SWID is added to the operation if one is not provided. Therefore if a SWID already exists in the browser it is not overridden. If the browser is yet to be configured then the new SWID will be used.