MAJOR/OPTIM/FEAT: Modified the project to use Go modules. Added additional features.
The config-go package uses viper to handle configuration. All configuration can be provided in files or environment variables. The vscode launch.json file that previously contained environment variables no longer does. The developer configuration is contained in the appsettings.dev.json and can be used without editing to get the demo running. Only the DNS step needs to be completed by the developer to get the demo running on a development machine.

Readme modified to remove reference to the deps.ps1 and deps.sh scripts that were used prior to Go modules. .gitignore has also been modified to remove the package directories that were previously created.

The build script on windows will now ignore some of the prebid files that are not needed in production.

Static file handler in the demo now works with exact paths. Needed to reference the prebid files.